+++ server/routes/userRoutes.js
@@ -63,7 +63,9 @@ router.route( '/revalidate/:email' )
 router.route( '/uploadProfilePicture' )
     .post( async ( req, res ) => {
 
-
+        console.log( "req.files: " + JSON.stringify( req.files ) );
+        console.log( "req.body: " + JSON.stringify( req.body ) );
+        
         if ( !req.files || Object.keys( req.files ).length === 0 ) {
             // no files uploaded
             await userController.saveProfileImage( req, res, req.session.authUser.email, 'profile-default.png' );
